progress implementing antigen test header view
