+++ src/xcode/ENA/ENA/Resources/Localization/de.lproj/Localizable.strings
@@ -665,7 +665,7 @@ Bei ausgeschalteter Hintergrundaktualisierung müssen Sie die App täglich aufru
 
 "ExposureSubmissionResult_CardSubTitle" = "IHR BEFUND";
 
-"ExposureSubmissionResult_CardSubTitle_Antigen" = "Befund";
+"ExposureSubmissionResult_CardSubTitle_Antigen" = "Ihr Schnelltest Befund";
 
 "ExposureSubmissionResult_CardTitle" = "SARS-CoV-2";
